+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Added support for direct station deauthentication
   - Added support for AirTag selection and spoofing
   - Added support for selecting and tracking Flipper Zero rssi
+  - Custom beacon SSID list management and spam
   
 - Commands
   - Added station selection capability to existing select command